chore: fix duplicate import
